Refatorado os modelos, disponibilizado o primeiro endpoint e iniciado test

+++ src/main/resources/db/migration/V02__valor_gasto.sql
@@ -1,6 +1,6 @@
 CREATE TABLE combustivel (
 	cod_combustivel INT PRIMARY KEY AUTO_INCREMENT,
-	preco_gasolina INT NOT NULL,
+	preco_gasolina DECIMAL NOT NULL,
 	total_km_percorrigo_C INT NOT NULL,
 	total_km_percorrigo_e INT NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
